strips brush rotation has major issues, when differnt brush width and height used the result is not expected, frequency increasse.

+++ TerrainToolsNew/Assets/TerrarinTools/Scripts/BrushData.cs
@@ -7,7 +7,7 @@ namespace TerrainTools {
         public float angle;
         public Vector2Int actualBrushSize;
 
-        public int stripCount;
+        public float stripCount;
         public float brushStrength;
         public float brushHeight;
